doc: Use CDN specific version for example instead of using latest
